Improve algorithm to migrate Jetpack sites
The previous code was too strict and would only do an exact match on the
XML-RPC endpoint.

This one will use the stored wp.com blog ID when available, and will fall back
to the XML-RPC endpoint, doing a case insensitive match of both HTTP and HTTPS
variants.

This is also closer to the algorithm in the Android app.