Use the loginURL to test for cookies.
If there's a mismatch between the login and target URL domains, we care more
about the login one. For instance, if we want to pre-auth with WordPress.com a
link to example.com, we'll never have cookies for example.com so it would have
authenticated every single time, just to get the WordPress.com cookies we
already had.