Removes shouldScrollToBottom from web view controller
This is a fun one. It was only used when the API failed because XML-RPC was
disabled, we'd load the settings page and scroll to the bottom, as that's where
the XML-RPC option was.

That option has been gone from settings for more than 5 years. Most comically,
it was removed just 6 days after our feature to show the settings page was
introduced.

https://github.com/wordpress-mobile/WordPress-iOS/commit/77771b264b53acae7ebfbe89a8bd421c00bff750
https://core.trac.wordpress.org/changeset/21804