Fixes a layout issue when returning from a search detail
I noticed that if I viewed a tag from a search result then navigated back
the table view's layout would have changed such that its content was offset
underneath the nav bar.
This appears to be related to setting a presentation context tho I wasn't able to
sort out *why* the layout was being changed.
Without setting a presentation context the search bar remains visible on the detail
screen.
This introduces a work around that preserves the tableview layout and
hides the search bar from the detail view.