Updates to also find an account by email address as well as username.
When an account is first created it might be created with an email address for its username.
This is eventually updated but there is a window of time where a second account with the
same username (but mismatched as an email address) can be created. Since usernames
and email addresses are both unique for accounts we can safely search for either and
prevent accidentally creating an extra account.