Fix connection flow redirect handling.
After a round of testing, it was reported that the flow was redirecting
to the plugin installation page right after selecting "Connect to
WordPress.com".

The problem was the current step switch was too specific, and it was
only clearing the site redirect if visiting an admin page other than
the expected ones (this was a workaround because poopy.life takes you
to wp-admin instead of the passed redirect_to URL).

This change will check if the detected step is any admin page and
prioritize using the redirect if there is a pending one.