Don’t try making screenshots without a valid login
If someone new to the project wants to work on screenshots, they can run the `WordPressScreenshotGeneration` test or use `fastlane snapshot`. However, if they haven’t set their WordPress.com username and password, it’ll fail with an unhelpful error:

>WordPressScreenshotGeneration.testGenerateScreenshots() failed: failed - Failed to find "Continue" Button after 30.0 seconds.

This is because by default the username and password are set to empty strings, which are technically valid code.

After this commit, if someone takes the same steps, it’ll instead tell them:
>WordPressScreenshotGeneration.swift:50:32: Editor placeholder in source file

This points them to the exact place they need to make a change in order to continue, which should hopefully ease the onboarding process.

Additionally, this change makes it possible to script the population of these values a little more easily – even if the variable names change, the tokens don't have to, and finding / replacing project-wide for these types of values becomes possible.