Add autosave for drafts
https://github.com/wordpress-mobile/WordPress-iOS/issues/10069

This is a first commit while I get myself familiar with the specs and the code base.
I've added a Debouncer class as one of the requirements suggests:
"Increase local autosave time/do autosave on every change (with some throttling)".

The plan is to use this Debouncer, so to not overload with network calls.