Auto saving drafts
When the user starts editing a post (recognized in textViewDidChange) we check to see if there are any changes to the post and if it has remote (so to save as local draft)
and then if so, we start a debuncer. The debunker will execute a call back in a given delay (500 ms in this case).
the call back is the same code we used for saving and updating a draft when the user tapped (save draft) from the secondary menu.