Fix loading slowness
When laying out attachment views, the `layoutAttachmentViewForAttachment` method would be called for each attachment. The resulted in nice, clean code, but it called `layoutManager.ensureLayout` for every layout attachment, which forces syncronous re-layout for each attachment. When there are many images in a post, this would take seconds of main thread work to do the layout.

Batching the layout work for attachments and only calling `ensureLayout` once removes a lot of unnecessary work.