Use the image size as specified by the <img> tag
Fixes a bug with gif images – using the `finalSize` works for non-gif images, but for gifs, the `efficientImageSize` method returns a height of zero.

Rather than rework `efficientImageSize`, it’s easier to just use the `proposedSize` for the attachment’s frame, and the gif can have variable quality within that – it may be downsampled within that frame, but it’ll always display at the right size.