Update the ZenDesk SDK to 2.3.0
What does this get us? From https://github.com/zendesk/zendesk_sdk_ios/releases:

Fix for image attachments layout issues in the RequestController, where images appear at varying dimensions in landscape.
Added Voice Over compatibility.
Fix for text cursor jumping to the end of text, when inserting unicode character, in the middle of the input field.
Fix for article failing to load on initial display, from a clean install.
Fix for tabBar blocking the inputField on iOS 9 and 10.
Deprecated the ZDKHelpCenterDelegate that displayed the human contact options,
use HelpCenterUiConfiguration and ArticleUiConfiguration instead.
Fix for navBar shadow missing after presenting Create Request screen.

The VoiceOver compatibility is important enough to justify the upgrade, IMHO. This change _may_ introduce deprecation warnings, but given that this is a point release and that we’re under no obligation to upgrade to the latest version yet, it seems worth it.