Reader: Check for finite value before setting offset
This PR addresses a crash in the reader detail that is not well understood.

We've received a number of reports of the app crashing when being opened. Crash reports
show that setContentOffset is being passed a non-finite value, leading to the crash. This
PR checks for a finite value before making the assignment, hopefully avoiding the crash.

We should be mindful of subsquent reports of layout issues with the reader detail after
opening the app. These could provide a clue as to the underlying cause.