Make NoticePresenter account for keyboard display
This is specifically useful for when a revision is loaded, a notice appears, THEN the keyboard appears, covering the notice.

For complete accounting of the keyboard the presenter would have to remember the keyboard state for when new notices appear, and place them in the correct position. However, that’s not required for this issue and was left out to reduce complexity.