Reader: Use WKWebView instead of the deprecated UIWebView for rich text embeds.
This refs an issue with video tags reported in 10930.
There seems to be an issue with UIWebView playing certain video formats.
That control is deprecated and WKWebView seems to support more formats so I've made that switch.

Bonus: WKWebView seems a bit smarter with its logic allowing us to get rid of some hacky logic UIWebView needed to display properly.