Reader: Set video tag height to auto.
Refs: #11078

This is a partial fix for the issues outlined in #11078. By setting the height to auto, we force the tag to preserve its aspect ratio.
With this change, a video tag that has a width and height defined via attributes will be correctly sized even tho we override the width to match the viewport.

This change also removes some extra whitespace around the top and left of the video by setting margins and padding to zero.