SiteCreation: Fix AddressCell height on first load
After some testing, this started happening in 4e6049a. That is a legitimate fix for the checkbox hovering over the labels. However it seems to be causing some issues on the self-sizing AddressCell. The issue only happens on the first table reload. This reverts the change on that.

To fix the checkbox hovering over the labels, I reverted the change in 6c84f94 and opted to increase the trailing constraint constant of the title label to accommodate the checkbox. This now works because the trailing constraint is tied to the Safe Area instead of the Content View.