Fix discard loosing all changes
When a post fails to save, it status was set to AbstractPostRemoteStatusFailed which should have
made post.hasNeverAttemptedToUpload return true hence return false in shouldRemovePostOnDismiss in discardChanges().
The problem was that the remoteStatus property was not really being saved. In order to save those change
we are merging the revisions back into the base post