Fix crash when syncing editor settings
When the mobile setting is not set, the saveRemoteEditorSettings will throw an error. This will call the failure handler but it doesn't return, so it will also call success a couple of lines later.

This means that the handler in BlogService will call dispatch_group_leave once too many times and the app will crash.

This change makes sure the error case returns early and the sync method doesn't call their callbacks more than once.