ReaderComments: Adjust comment service save behavior. Call failure from main thread.
Ensures a failure block is called from the main thread.

Previously contexts were saved by calling saveContextAndWait in order for the merge method to return a value.
This, unfortunately didn't handle derived contexts as intended.
This calls saveContext:withCompletionBlock to properly save a derived context and reworks how previous return value is provided to the caller.