Show “Changes will be uploaded” in Post List
This will be shown for posts with remote that have local changes and are confirmed to be automatically uploaded when the device is back online.

This also reverts 4ce88f3. Now, we will show “Draft will be uploaded…” for local drafts instead of ”Local changes” only. This was discussed with Chris and it seems much clearer for the user.