Automatically upload published and draft posts (#12466)
* Include posts with remote in auto-upload

* Use a different message for auto-uploaded drafts

Instead of just “Upload failed”, we’ll show “Draft will be uploaded…” instead.

* Update PostCardStatusViewModelTests

* Update PostCardCellTests for draft and published

* Change local draft to just show “Local changes”

The message “Draft will not be uploaded” is probably not necessary for it since it’s not cancelaeable anyway.

* PostNoticeViewModel: Promote autoUploadInteractor

* PostNoticeViewModel: Move failure titles to enum

* Show “Draft will be uploaded…” notice

This will affect drafts with remote only.

* PostCardCell: Add test for local draft message

* Show “Changes will be uploaded” in Post List

This will be shown for posts with remote that have local changes and are confirmed to be automatically uploaded when the device is back online.

This also reverts 4ce88f3. Now, we will show “Draft will be uploaded…” for local drafts instead of ”Local changes” only. This was discussed with Chris and it seems much clearer for the user.

* Show “Draft will be uploaded” in notice

And for posts with remote, always use “Changes will be uploaded…” as the notice message.

* Add additional tests in PostNoticeViewModelTests

* Add PostCardStatusViewModel.generateFailedStatus()

This is just a helper method to make things a bit readable. It should only be used by `private var status: String?`

* Remove unreliable post.remoteStatus assertion

This should have been a `post.isFailed` check but it’s better to just rely on the `guard` instead.