Update PostListFilter to use statusAfterSync
This change makes it so that posts do not immediately move to the corresponding tab if their statuses are changed. For example, if an existing published post is transitioned to a draft, that post will stay in the Published list until it is successfully uploaded. Once it is uploaded, it will be moved to the Drafts list.