Fix stuck “uploading” for self-hosted autosave
There is really no auto-save for self-hosted sites but we call PostService anyway. That’s how the previewing currently works.

This specific change is to fix the `remoteStatus` which gets stuck in `.uploading` when we call the failure block of `PostService.autoSave()`. The stuck `.uploading` is highly visible in the Post List.