Add MediaService.Factory
This is for testability. We will be mocking MediaService later.

Instead of trying to analyze whether we can get away with having just one `MediaService` inside `MediaCoordinator`, which requires quite a lot of regression testing, I’m opting to keep the `MediaService` object instantiation but hiding it inside a factory.