Use mainContext in PostCoordinatorTests
This fixes a few failures in the class if Core Data concurrency debug (`-com.apple.CoreData.ConcurrencyDebug 1`) is turned on.

Also fixed here is a concurrency issue in `testTracksAutoUploadPostInvoked`. The placement of `action` is incorrect because the `interactor.autoUploadAction()` would have returned a different value by the time that `resume()` finishes.