Auto-upload unconfirmed posts as drafts
For self-hosted sites posts that were unconfirmed for auto-uploading, we will still upload them but as drafts.

I implemented as a new action in `PostAutoUploadInteractor` rather than adding another logic for self-hosted sites in `PostService.autoSave`. The `autoSave` path can get tricky because it is also used in previewing. Enabling upload as draft in `autoSave` would cause previews to be confusing for users. Consider this scenario:

1. Create a post. Set its status as published Post Settings.
2. Preview — the post will be auto-uploaded as a draft.
3. Preview again — the post will no longer be auto-uploaded as a draft. We cannot upload it as `published` either because the user is still writing the post. So in this case, preview will just show a “Cannot preview“ error.

With this, I opted to have a separate action and use `PostService.uploadAction`.