DomainServiceTests: Increase timeout
Every so often, I get unit test failures in CircleCI because of `DomainServiceTests`. I looked at one of the successful results and found that one of tests is pretty close to 1 sec.

```
DomainsServiceTests
    ✓ testDomainServiceHandlesTwoNewDomains (0.868 seconds)
    ✓ testDomainServiceParsesAllDomainTypes (0.048 seconds)
    ✓ testDomainServiceParsesPrimaryDomain (0.044 seconds)
    ✓ testDomainServiceRemovesOldDomains (0.051 seconds)
    ✓ testDomainServiceUpdatesExistingDomains (0.044 seconds)
```

Something might have changed somewhere that resulted in this test to be not as fast as we thought it was.