Create edit revision before loading Gutenberg
When a post is edited, the editor is loaded with the original post and
then a revision is made. This order seems incorrect - the revision
should be made _before_ being loaded into the editor.

This change will allow autosave content that will be conditionally added
to a post revision to be loaded into the editor.