Remove unintended change from release notes
I noticed that there was a line in release notes that shouldn't be there
(it duplicated an existing line):

> * Block editor: Images from Image Block can now be previewed full screen by tapping on them.

This line was introduced in 6716fc84ecc and moved in ffdbfdb03877. It
looks like a merge or conflict was not properly resolved at some point
on my feature branch resulting in the move commit not being cleanly
applied (line was added to a new line position but not deleted from its
old position).

This looks like the only inadvertent change - all else looks fine.