DatePicker: Always show close button for a11y
This shows the close button if VoiceOver or Switch Control is enabled. I wish we could turn it on for Voice Control too but there does not seem to be an API for that yet.

Ideally, we should also have a transparent button on the top part of the `HalfScreenPresentationController`.