Set up Yarn to fetch Danger and TypeScript
In https://github.com/wordpress-mobile/WordPress-iOS/pull/13423 we set
up Danger to run via GitHub workflows and perform checks on the PR
labels and iOS code consistency.

The setup used Danger via GitHub Action
(https://github.com/marketplace/actions/danger-js-action). It's
convenient to use the Action because we don't have to manage Danger
locally. On the other hand, the action takes more than 1 minute to run.

By switching to a local setup, we can run the same Danger checks in ~20
seconds.

Something to note about the setup is that we fetch TypeScript as well as
Danger. That's because our Dangerfiles are written in TypeScript
(https://github.com/Automattic/peril-settings/tree/5d22258ca46c3649f2cb863cfac2d16f12c4cba0),
so Danger needs the TypeScript transpiler to be in the node_modules/
folder in order to read them.