Fix State Restoration issue + feature flag check
This fixes an issue where state restoration would fail on Blog Details View Controller because the restoration properties were not set.
Defaults to hiding the create button instead of leaving it shown.
Also adds a check for the CreateButton feature flag to prevent a possible crash if that feature flag is disabled.