Add support for animated GIFs in Gutenberg.
When we implemented our custom ImageLoader for RN, to support private
sites, we accidentally remove support for playing animated GIFs while
inside GB editor.
This PR changes the loading system of images to use the
RCTAnimated class when a GIF is being download.

One of the side effects of that change is that if now the featured
image on the post/page is an animated GIF we show the animation.