Support dark mode on pre-blocks in post history
When viewing the visual preview of a post revision, the background color of pre-based blocks (e.g. verse, preformatted) did not adapt to Dark Mode.
This is fixed by setting the `preBackgroundColor` of the Aztec TextView being used. A color of neutral (shade 5) is used because it's what the Classic editor uses.