Remove GitHub Action workflow to run Danger on PRs
The plan was to use this stack instead of Peril on Heroku, but we bumped
into a few too many integration issues.

I'll be keeping an eye on Danger and revisit the feasibility of the
project periodically. I have no doubt that as Danger evolves, we'll be
able to adopt it without the tradeoffs we would have had to made at this
point in time.