Rename Contact Us to Contact Support
To clarify to app users that the Contact button is to contact WordPress.com support and not a place to leave contact details for _their_ site visitors to contact their the site, we clarified the button title to "Contact Support".

Note: The Zendesk chat screen title should also be changed, but it requires manually adding a string to the localized string files, which is incompatible with our approach of auto-generating string files from NSLocalizedString usage in the codebase.