Log error if preview auth cookie is not found
When the user previews a post or page, the app fetches an auth cookie from the server. It appears that there are some instances where the server sends back an empty auth cookie, so this PR adds a check there and logs an error if the auth cookie is missing from the server's response.