Fix drafts with revisions disappearing upon update
Changes predicate to allow original posts with conflicts
and revisions in other tabs to be displayed.

This can happen when a local draft has been autosaved then that same post
is published remotely. The published post is not a revision but
should be displayed as the revision is in the draft tab.