Prevent accidental changes to App Store Connect
Without the `readonly` option set to `true`, in some cases `match` may attempt to make changes to App Store Connect Certificates and Profiles. This can cause other code signing problems, so we want to avoid it.

A future PR will add the ability to regenerate these profiles yearly as needed.