Set `hasRevision` only if remote post is new
This was previously done by disabling `updatePost()` when the post isn't new.
However we need to update when the actual remote post isn't new but does contain
a remote autosave.

This also has to be done before any copying of the remote post
to the original post takes place so has been moved to the top
of the function.