Disable block editor informative popup
We'll soon be removing the ability to switch to Aztec, so in preparation for that, this PR disables the popup that informs users that the app has switched them back over to Gutenberg.

We will remove the dead code in a future PR (issue is https://github.com/wordpress-mobile/gutenberg-mobile/issues/2901).