Hide Xpost UI on unsupported sites
Fixes https://github.com/wordpress-mobile/gutenberg-mobile/issues/2942

Previously, sites that don't support Xposts (e.g. Simple WordPress.com sites) showed the suggestions UI when the user typed a `+` character. This PR fixes this by disabling the suggestions UI as soon as the Xpost prefetch request returns a result from the network that indicates that Xposts are not available.

If after the app determines that a site has Xposts disabled, the site then becomes Xpost-enabled remotely, the app will be able to pick up this change only once the app is restarted (that is, when the `SiteSuggestionService` singleton is cleared from memory).