Fix Xpost insertion
Xposts added from WordPress for iOS were not triggering an Xpost (cross-post) to be created on the site that was cross-posted to. This meant you could add an Xpost from the WPiOS app, but it wouldn't actually make a cross-post.

The reason was that the app was not inserting the destination site's subdomain (e.g. `+testsiteb`) but instead it was inserting the site's title (e.g. '+Test Site B'), which is not interpreted by the backend as a trigger to make a cross-post.

This PR fixes that by inserting the site's subdomain instead of its title.