Don't check editor pref on new Simple site posts
Before this commit, when creating a new post the app would check the site's mobile editor preference (which is different to their web editor preference) and open the corresponding editor (Block editor or Classic editor).

With this commit we now do not look at the preference when the user creates a new post and we will now always use the Block editor in this scenario).