Fix suggestion crash on offline devices
The app crashed when the device was offline and the suggestions code (which handles Mentions and Cross-posts) returned a nil result. This result was interpreted by the SuggestionsTableView code as "suggestions haven't been fetched yet". The code, which has logic for fetching suggestions tied to the suggestions data source (a property), then entered an infinite loop of reloading the suggestions list.

The fix here is to undo an architectural decision made long ago to tie the fetching of suggestions to the suggestions data source. I've used didMoveToSuperview on the suggestions view as the trigger to fetch suggestions. This method, somewhat similar to UIViewController's viewDidLoad, is called once each time the suggestions UI is shown.