Fix Mentions/Xpost crash
The crash happens when the device is offline and the suggestions code (which handles Mentions and Xposts) returns a `nil` result. This result is interpreted by the `SuggestionsTableView` code as "suggestions haven't been fetched yet". The code there, which has logic for fetching suggestions tied to the suggestions data source (a property), then gets in an infinite loop of fetching, failing and retrying until a stack overflow exception occurs.