Fix broken blog lookup from Objective C
Prior to this commit, Blog lookup from Objective C could crash if passed a `nil` NSNumber, because the Objective-C to Swift bridge doesn’t have enough type information to complain about it being an invalid operation.

To fix it, we convert the `NSNumber` object to a primitive – if the `NSNumber` is nil, messaging `int64Value` will return `0` which we can look up using the existing code path.