Dismiss alert in UI tests
The new classic editor deprecation alert in Aztec (classic editor) is shown only once and was making all UI tests fail on the first try. CircleCI then ran tests a second time and because the alert is not shown again (a flag is persisted to the device to ensure this), tests passed the second time around.