New 'GutenbergFeaturedImageHelper' class
With this commit, the setFeaturedImage() function is moved away from 'GutenbergMediaInserterHelper' and to a new  'GutenbergFeaturedImageHelper' class. The code serves a slightly separate purpose to the other code in 'GutenbergMediaInserterHelper', so the aim of moving it to a new class is to improve overall readability.

References/calls to the old class/function are also updated in GutenbergViewController.swift as part of this commit, to ensure the 'set featured' functionality continues to work.