Refactor if/else statement for readability
With this commit, the if/else statement is refactored to a series of single if statements, with accompanying comments. The goal is to make the code more readable and for it to be clearer what the purpose of each combination is. It also adds an explicit if statement for any cases where the media ID may be equal to the featured image's ID.

This brings the code inline to that used for the Android side of this feature. See: https://github.com/wordpress-mobile/WordPress-Android/pull/14503/commits/fe93e0b88405bd4240c6263769ed56fae72f14d0