Introduce 'mediaIdNoFeaturedImageSet' constant
With this commit, a 'mediaIdNoFeaturedImageSet' constant is introduced and used in checks where a featured image is being removed.

The purpose behind introducing a constant is to improve readability and make it easier to change in the future, if needed.

This brings the code inline with the Android side of the functionality. See: https://github.com/wordpress-mobile/WordPress-Android/pull/14503/commits/9630c2955da873dbbd033849a6d1494876bbb582