Unwrap 'featuredImageId' to prevent crash when nil
When the featured image's ID was nil (that is, when there was no featured image) and an attempt was made to set a featured image, the app crashed with the following error: "force unwrapped a nil value"

The crash is fixed with this commit, by checking whether the 'featuredImageId' is nil in the first check within the 'gutenbergDidRequestToSetFeaturedImage' function.

With this commit, a new 'featuredImageId' has also been created, as its value is now used in three places in the code block.