Reposition 'mediaIdNoFeaturedImageSet' constant
With this commit, the constant has been moved from the top of the file to inside the 'gutenbergDidRequestToSetFeaturedImage' function where it's used.

The constant was repositioned in a bid to improve the overall readability of the code and to follow the logic I'm seeing elsewhere in the class (I'm not seeing a pattern of constants at the top of the file).