Edit descriptions of strings for clarity
The existing descriptions accompanying strings have been edited with this commit, with a view to make them as clear as possbile for translators.

Two descriptions have also been added to strings that didn't have them ("replace" and "cancel").

Finally, the "Replace current?" string have been updated to "Replace" in a bid to make it more translateable.