Refactor logic around dismissing controller
It's necessary to dismiss then present controller at the moment due to an issue with BottomSheet. See hee for further discussion: https://github.com/wordpress-mobile/WordPress-iOS/pull/16426/files#r643544481

With this commit, the logic around dismissing the controller is refactored to make use of the completion handler.

An if/else statment is used to account for any cases where a controller isn't present (i.e. if the issue with BottomSheet is fixed in the future).