Use launch environment instead of bundle to read WireMock settings
Using the bundle broke once I moved the class to a different target.
Given that all those values were hardcoded anyways, I decided to
hardcode them in the new implementation, too, and to leverage the launch
environment variables (which can be configured in the scheme or test
plan used to run the tests) to allow users to use a custom host and port
for WireMock.