Merge develop into release/17.8
`develop` as at 5100e7983b21c6b33204867c86c4aff194cb0ed2.

Got conflicts on:

- `Podfile`
- `Podfile.lock`
- `WordPress/WordPressTest/BlockEditorSettingsServiceTests.swift`

I solved the `Podfile*` picking up the version from `develop` because
it's the latest.

I solved the test issue picking the changes from the release branch,
because they were needed for the hotfix code.

The tests are failing for me locally in that test class with:

```
*** Terminating app due to uncaught exception 'NSInternalInconsistencyException', reason: 'API violation - multiple calls made to -[XCTestExpectation fulfill] for Theme should be successfully fetched.'
terminating with uncaught exception of type NSException
```