Merge Xcode image and environment definition in Buildkite pipeline
See suggestion from @AliSoftware here:
https://github.com/wordpress-mobile/WordPress-iOS/pull/16929#discussion_r677323734

I liked the extracted Xcode version because it made it easier to find
it, IMHO, but I'll admit it was a bit over-engineered, in particular
given we don't have any need to use it on its own right now.

I chose the approach in this commit because I like how it forces each
step to explicitly define the `env` key. I feel doing so keeps the step
syntax more explicit. I our pipeline files grow in size, a reader won't
have to jump to the definition of the anchor to know it's used for the
`env` key, which they would have had to do with the `<<: *common_env`
approach.

I also decided to use underscores instead of the suggested dashes to
keep it in line with the other anchor already defined (`use_plugin`) and
the Buildkite spec (
https://github.com/buildkite/pipeline-schema/blob/f7a0894074d194bcf19eec5411fec0528f7f4180/schema.json
) which uses underscores, e.g. in `artifacts_paths`.