Modify updateLikes method to accept String instead of LikeUser instances
This change will help simplify the implementation later on when reacting to like button actions from ReaderDetailToolbar. Previously, if the method requires LikeUser, then we'd have to unnecessarily create `LikeUser` objects based on `WPAccount`.

If the method accepts URL strings instead, we can just simply pass in the value of `avatarUrl` from `WPAccount` without converting it to a `LikeUser` instance.