Featured Image dismisses view controller if not already dismissing
We now trigger the Featured Image flow whenever an Image block's media is
replaced. A race condition between the Media Picker dismissing itself
and the replaced media triggering the Featured Image flow caused the
entire editor to be dismissed. To avoid erroneously dismissing the
editor itself, we should not dismiss the `presentedViewController` if it
is already actively dismissing.