Set the webview back to opaque after content has finished loading.
The web view needs to be translucent initially, because otherwise the web view might show a short, white flash. This is noticeable in dark color schemes, and a tad annoying. On the other hand, a translucent webview causes scroll indicator to be invisible, regardless of the indicator style.

Therefore, the solution is to set the webview to translucent initially, and set it back to opaque once the content has finished loading.