Temporarily remove aggregate target for credentials codegen in Jetpack
We noticed that Google Login was broken — resulting in the app to crash
– in Jetpack.

My diagnosis is that was because the codegen script receiving a
`BUILD_SCHEME = WordPress` value, instead of `Jetpack`. The reason, I
think, is that the aggregate target has it's own set of build settings
and doesn't use the Jetpack ones, despite the build being set with the
aggregate target as a dependency for Jetpack.

To resolve this in the short term, I have rolled back to running the
codegen script as a Build Phase Run Script as part of the Jetpack
target, removing the dependency on the aggregate target.

There might be better ways to do this in the long run, but for the
purpose of fixing this in the code freeze branch or to ship a hotfix,
this was the fastest thing I could come up with.