Make pre-publish sheet header opaque
Fixes https://github.com/wordpress-mobile/WordPress-iOS/issues/17143

The pre-publish header was transparent when it should have been opaque. This was noticable if the user scrolled the sheet contents up under the (transparent) header.

Instead of simply setting the header's background color, this fix works by setting the header's background view and settings _its_ background color. This approach is described in https://stackoverflow.com/a/25588828/1305067 and is the only known way to set the background color of a `UITableViewHeaderFooterView` (which is what this table header is).