Remove dedicated release note entry for 18.6.1
See conversation here:
https://github.com/wordpress-mobile/WordPress-iOS/pull/17483#discussion_r751494058

My rationale for adding it was to track the reason for 18.6.1 in
`RELEASE-NOTES.txt`.

My choice raised a few questions and doesn't align with what other apps
do. For the sake of simplicity, it might be best to revert.

We have internal documentation for each release, GitHub PRs, and the
possibility to use Git to diff between tags (e.g. `git diff 18.6
18.6.1`). That information won't be lost and only a few extra keystrokes
away.