Do not use random values in tests without having a seed
Random values in tests are great to catch weird behavior and input edge
cases that the test author might not have considered at the time of
writing.

But, one needs a way to repeat the test with the same input in order to
reproduce the issue. Otherwise, we risk getting flaky tests with no way
of knowing why.

Without a configurable seed from which to generate random values, random
input can be dangerous for the consistency of the test suite.

Another reason I'm making this change is because the Simulator takes
a long time to type such long inputs one character at a time.