Move AppLocalizedString.swift helper to shared dir
This file was previously in the directory associated with the `WordPressStatsWidgets` target.

But now that I plan to use this helper for other App Extension / Widget targets, it seems better suited to move it under `WordPress/Utility`.
For now this commit only moves it around, next commits will add that file to more targets and start using it elsewhere.