Make `Double+Stats.swift` use `AppLocalizedString`
This file is used not only by the app, but is also build as part of the various Widget targets

Note: searching each swift file that is built as part of by `WordPressStatsWidgets` (iOS14) for their use of `NSLocalizedString` revealed that this was the only one which we missed during transition of the WordPressStatsWidgets target previously.

Other widget targets have more files which need the same transition, those will be done in a subsequent commit.