Make SwiftLint Build Phase be compatible with rbenv
By sourcing `~/.zshrc` if it exists, so that the `rbenv init` and friends are evaluated even in the context of the Xcode shell.

Otherwise, the script build phase will either not find `bundler` (if it's not installed system-wide), or find bundler but then error with "Bundler::GemNotFound: Could not find rake" because it would not use the bundler from the current rbenv'd ruby and not find the related gems like rake installed by it.