Clean up i18n function and update exclusing list for genstrings
`NSLocalizedString` is a macro only meant to be used with string literals and for cases when you want `genstrings` to extract the values.
For any other method needing a dynamic localized string key (determined by something that is not a String literal) and/or that is not meant to be parsed by `genstrings` (like here for unit tests), the raw `Bundle.localizedString` method should be called directly instead.