Add new `BUILD_NUMBER` config to use in place of `VERSION_LONG`
The App Store build for Jetpack was uploaded with build number 20. This
means that App Store Connect will reject any build with number less than
21, including the current value of 18.x.y.z.

This happened already in Simplenote macOS and we updated it and the
release toolkit to support a new, plain integer build number under the
config key `BUILD_NUMBER`.

Note that this is used **only for Release builds**. Debug and
Internal/Alpha build will keep using the `VERSION_LONG` build number to
keep the more granular and calendar based numbering.