17624 Prevent double-push of edit username screen
The edit username screen on the SignupEpilogue screen was sometimes pushed twice when a user tapped on the username text field.

This navigation was triggered in the implementation of `UITextFieldDelegate.textFieldShouldBeginEditing`, which was under certain circumstances, called twice by the system. That delegate method is supposed to answer the question "should editing be allowed?" with a Bool, so navigating here could be considered a side-effect, which makes the delegate method unsafe to call more than once.

This change moves the responsibility for navigation to the TableViewController, in the `didSelectRowAtIndex` method.

The `UITextField` for username also needed to be disabled, to make it easier to tap the row. Without this change, the user would have to tap on the "Username:" label, or the disclosure indicator, and taps on the text field would be ignored. Prior to this commit, users would have to tap on the username value itself, taps elsewhere on the row were ignored.